Fix several more bugs in 0.5.2.

+++ src/externs.h
@@ -1297,7 +1297,7 @@ extern void grid_queue_destroy(grid_queue_type *q);
 extern bool grid_queue_push(grid_queue_type *q, byte y, byte x);
 extern void grid_queue_pop(grid_queue_type *q);
 extern int pick_random_item(int chance_values[], int max);
-extern errr next_line_to_number(ang_file *fff, int *dest);
+extern errr next_line_to_number(ang_file *fff, byte *dest);
